Refactor check-fjord script and e2e test for cleaner separation
